refactor: abstract out Keplr wallet sync logic
